Moved some things around and updated the scheduler and the timer creation

+++ DayZServerManager.Server/Classes/SerializationClasses/ManagerClasses/ManagerConfigClasses/CustomMessage.cs
@@ -4,14 +4,14 @@
     {
         public int Id { get; set; }
         public bool IsTimeOfDay { get; set; }
-        public Dictionary<string, double> WaitTime { get; set; }
-        public Dictionary<string, double> Interval { get; set; }
+        public Dictionary<string, int> WaitTime { get; set; }
+        public Dictionary<string, int> Interval { get; set; }
         public string Title { get; set; }
         public string Message { get; set; }
         public string Icon { get; set; }
         public string Color { get; set; }
 
-        public CustomMessage(int id, bool isTimeOfDay, Dictionary<string, double> waitTime, Dictionary<string, double> interval, string title, string message, string icon, string color)
+        public CustomMessage(int id, bool isTimeOfDay, Dictionary<string, int> waitTime, Dictionary<string, int> interval, string title, string message, string icon, string color)
         {
             Id = id;
             IsTimeOfDay = isTimeOfDay;
